# KAFKA-16284: Fix performance regression in RocksDB