A performance regression introduced in commit 5bc3aa428067dff1f2b9075ff5d1351fb05d4b10 reduces the write performance in RocksDB by ~3x. The bug is that we fail to pass the `WriteOptions` that disable the write-ahead log into the DB accessor.

For testing, the time to write 10 times 1 Million records into one RocksDB each were measured:

 * Before 5bc3aa428067dff1f2b9075ff5d1351fb05d4b10: 7954ms, 12933ms
 * After 5bc3aa428067dff1f2b9075ff5d1351fb05d4b10: 30345ms, 31992ms
 * After 5bc3aa428067dff1f2b9075ff5d1351fb05d4b10 with this fix: 8040ms, 10563ms
 * On current trunk with this fix:  9508ms, 10441ms
 
### Committer Checklist (excluded from commit message)
- [ ] Verify design and implementation 
- [ ] Verify test coverage and CI build status
- [ ] Verify documentation (including upgrade notes)